requirements.txt
Django==2.2.16
psycopg2-binary==2.8.6
python-decouple==3.3
djangorestframework==3.12.1
